% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_cran_downloads.R
\name{adj_cran_downloads}
\alias{adj_cran_downloads}
\title{Adjusted Cran Downloads}
\usage{
adj_cran_downloads(
  packages,
  when,
  from,
  to,
  remove_update_spikes = TRUE,
  number_to_compare = 100,
  value_quantile = NULL
)
}
\arguments{
\item{packages}{A character vector, the packages to query.}

\item{when}{\code{last-day}, \code{last-week} or \code{last-month}.
If this is given, then \code{from} and \code{to} are ignored.}

\item{from}{Start date, in \code{yyyy-mm-dd} format, or
\code{last-day}. It is ignored if \code{when} is given.}

\item{to}{End date, in \code{yyyy-mm-dd} format, or
\code{last-day}. It is ignored if \code{when} is given.}

\item{remove_update_spikes}{Default `TRUE`. Removes the spike in downloads due to automatic re-downloads
and CRAN mirrors associated with a package update. This replaces the number of downloads on an update day
and the day following with the median download value across the measurement period. You can set this to `FALSE` if you do
not see spikes in downloads on days before or after you update to accurately capture downloads during those days.}

\item{number_to_compare}{Default `100`. The number of random packages to inspect to determine the minimum number of downloads in any single day.}

\item{value_quantile}{Default `NULL`. By default, the function uses the minimum number of downloads. Setting a number here
(between 0 and 1) instead returns the nth lowest number of downloads, where n = `value_quantile` * `number_to_compare`.}
}
\value{
A data frame of downloads and total downloads by package.
  \item{\code{date}}{Day of the downloads, it is a Date object.}
  \item{\code{package}}{The package. This column is missing if
    all packages were queried.}
  \item{\code{count}}{Raw download count.}
  \item{\code{total_downloads}}{Raw total number of downloads from beginning of measurement period.}
  \item{\code{mindownloads}}{Estimated CRAN-wide automated downloads for that day.}
  \item{\code{updateday}}{TRUE if the package was updated that day, FALSE otherwise.}
  \item{\code{adjusted_downloads}}{Daily download count, adjusted.}
  \item{\code{adjusted_total_downloads}}{Adjusted total number of downloads from beginning of measurement period.}
}
\description{
Removes daily shared downloads amongst CRAN packages and re-download/CRAN mirror spikes associated with
package updates. The function samples a number of packages from the CRAN and finds the minimum
}
\examples{

\dontrun{adj_cran_downloads("skpr",from="2017-08-15",to="2017-11-15")
adj_cran_downloads("skpr",when="last-month")
adj_cran_downloads(c("skpr","AlgDesign"),when="last-month")}
}
